pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aegis"
version = "1.0.0"
description = "Agentic Exploit & Guardrail Investigation Suite — security testing framework for agentic AI systems"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "langchain==0.3.27",
    "langchain-community==0.3.31",
    "langgraph==0.2.76",
    "fastmcp==2.14.5",
    "chromadb==1.5.0",
    "pydantic==2.12.5",
    "typer==0.23.1",
    "jinja2==3.1.6",
    "pyyaml==6.0.3",
    "deepeval==3.8.4",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "jsonschema==4.26.0",
    "pyright>=1.1.408",
    "ruff==0.15.1",
]
redteam = [
    "pyrit==0.11.0",
]
dashboard = [
    "streamlit>=1.30,<2.0",
    "plotly>=5.18",
    "pandas>=2.1",
]

[project.scripts]
aegis = "aegis.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["aegis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -q -p no:deepeval --ignore=tests/test_orchestrator.py"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["aegis"]

[tool.pyright]
include = ["aegis", "dashboard"]
exclude = [
    "**/__pycache__",
    ".venv",
    ".venv.broken-20260307",
    ".codex",
    "output",
    "reports",
    "tmp",
]
pythonVersion = "3.11"
pythonPlatform = "Linux"
typeCheckingMode = "basic"
extraPaths = ["."]
reportMissingTypeStubs = "none"

[tool.coverage.run]
source = ["aegis"]
omit = [
    "tests/*",
    "aegis/testbed/agent.py",
    "aegis/testbed/mcp_servers/*",
    "aegis/testbed/rag_pipeline.py",
    "aegis/orchestrator.py",
]

[tool.coverage.report]
fail_under = 80

[dependency-groups]
dev = [
    "coverage==7.13.4",
    "jsonschema==4.26.0",
    "pyright>=1.1.408",
    "pytest-cov==7.0.0",
    "ruff==0.15.1",
]
